+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Each log message now shows a detailed view if you click on them (if you have the permissions)
 - Enums now use a native Zirconium enum type, rather than a hacky userdata.
 - Added the `players` (array of players), `players?`, `range` (in Zirconium e.g. `1..10`) and `defined` built-in argument types.
+- Added union arguments to `ZirconFunction`s
 - Added variadic arguments to `ZirconFunction`s
 - History now works properly
 - Auto-focus is back, but now as an option to the client `Init` function